Added 'CompatiblePSEditions' to module manifest
The module already requires a language version of 6 or greater, however, this doesn't block the installation of the module on Windows PowerShell (i.e. v5.1) - it just prevents it from being imported and makes it invisible to 'Get-Module', which doesn't help diagnosing this scenario.

This change should ensure that attempts to install the module on older versions of PowerShell are blocked.

Technically this a breaking change.