Resolve issue with logic that removes the temporary network access rules for Key Vaults.
The behaviour was being masked by a null reference bug.

We need to handle the fact that Key Vaults stores rules for single IP addresses in CIDR notation (i.e. /32) even when not specified in that form (unlike Storage Accounts) which require single addresses to be specified without the CIDR notation.